+++ CHANGELOG.md
@@ -27,11 +27,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Upgraded okhttp3 to v4.10.0 (this is a major version upgrade from v3 to v4)
 - Upgraded gson dependency to v2.8.9
 - Upgraded androidx.appcompat dependency to v1.6.1
-- Upgraded example apps to latest bitmovin player (v3.29.0) and latest exoplayer (v2.18.3)
+- Upgraded example apps to latest bitmovin player (v3.33.0) and latest exoplayer (v2.18.3)
 - Deprecated obsolete PlayerType config in AnalyticsConfig
 - use `System.getProperty("http.agent")` for user agent tracking
 - Made all bitmovin and exoplayer collector classes that are not part of the public API internal
-- Stopped sending samples once player is released
+- Stopped sending samples once player is released and collector is still attached
 
 ### Fixed
 - [internal] reporting of videoTimeStart and videoTimeEnd for audioTrackChange, subtitleChange, qualityChange